api_yamdb/requirements.txt
asgiref==3.2.10
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==2.2.16
django-debug-toolbar==3.2.2
django-debug-toolbar==3.2.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
flake8==4.0.1
gunicorn==20.0.4
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
packaging==21.2
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.2
pytz==2020.1
requests==2.26.0
sort-requirements==1.3.0
sqlparse==0.3.1
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7